ovl: warn about orphan metacopy
From 0a8d0b64dd6acfbc9e9b79022654bbe1ade4a29a Mon Sep 17 00:00:00 2001
Git-commit: 0a8d0b64dd6acfbc9e9b79022654bbe1ade4a29a
Patch-mainline: v5.11-rc1
References: bsn#1

When the lower file of a metacopy is inaccessible, -EIO is returned.  For
users not familiar with overlayfs internals, such as myself, the meaning of
this error may not be apparent or easy to determine, since the (metacopy)
file is present and open/stat succeed when accessed outside of the overlay.

Add a rate-limited warning for orphan metacopy to give users a hint when
investigating such errors.

Link: https://lore.kernel.org/linux-unionfs/CAOQ4uxi23Zsmfb4rCed1n=On0NNA5KZD74jjjeyz+et32sk-gg@mail.gmail.com/
Signed-off-by: Kevin Locke <kevin@kevinlocke.name>
Signed-off-by: Miklos Szeredi <mszeredi@redhat.com>
Signed-off-by: Guoqing Jiang <guoqing.jiang@suse.com>

Patch-name: patches.mainline/v5.11-rc1-ovl-warn-about-orphan-metacopy.patch