tcp: propagate MPTCP skb extensions on xmit splits
Git-commit: 5a369ca64364b49caff424d2f988901bc7658b6d
Message-Id: <5a369ca64364b49caff424d2f988901bc7658b6d.1637075535.git.geliang.tang@suse.com>
References: bsn#2
Patch-mainline: v5.11-rc1

When the TCP stack splits a packet on the write queue, the tail
half currently lose the associated skb extensions, and will not
carry the DSM on the wire.

The above does not cause functional problems and is allowed by
the RFC, but interact badly with GRO and RX coalescing, as possible
candidates for aggregation will carry different TCP options.

This change tries to improve the MPTCP behavior, propagating the
skb extensions on split.

Additionally, we must prevent the MPTCP stack from updating the
mapping after the split occur: that will both violate the RFC and
fool the reader.

Signed-off-by: Paolo Abeni <pabeni@redhat.com>
Signed-off-by: Mat Martineau <mathew.j.martineau@linux.intel.com>
Signed-off-by: Jakub Kicinski <kuba@kernel.org>
Signed-off-by: Geliang Tang <geliang.tang@suse.com>

Patch-name: patches.suse/v5.11-rc1-tcp-propagate-MPTCP-skb-extensions-on-xmit-splits.patch